Domain Classes
REQUIREMENTS: For each Gig we need to capture WHO, WHERE, WHEN & what
the genre is. We will start with the WHO part!

We need to associate the Gig.cs class with another class that represents
a user in GigHub. That class already exists by default if you check in
IdentityModels.cs, it's called  ApplicationUser class and we call it
Artist!

WHEN part is next and we use the DateTime class for this

WHERE is this going to happen? We need to capture name of venue so we
use string called Venue

WHAT genre is it?  We will need another class with an Id and Name. Moved
this into it's own separate file to keep code clean - shortcut with
ReSharper ALT+Enter, make sure curser is on class name