Migration to Update Database
We created our domain classes, current;ly Gig and Genre, but not
anything to keep track of upcoming gigs or artists a user is following
because they are part of future Use Cases, doing iterative development -
Simple and Lean

So skip and do migration to update database

PROBLEM
add-migration CreatingGigTable - This initially did  not create tables
because i created a domain class but there was no reference to it
anywhere so EntityFramework does not know the existance of this class
the domain model
SOLUTION: Add a reference by going to Solution Explorer -> Models ->
IdentityModels.cs Then add a DbSet of type Gig so EntityFramework knows
about our Gig.cs class. Gig already has a reference to the Genre class
so EntityFramework will be able to find it but if you want to query the
list of genres we need to add a sepate DbSet of genres

Moved ApplicationDbContext from IdentityModels.cs into it's own separate
file to keep the code clean - I bulit the program a few times to make
sure there are no errors!

Renamed IdentityModels.cs to ApplicationUser.cs to match type name!