Add-Migration OverrideConventionsForGigsAndGenres
PROBLEM: Strings are Nullable in C# so EnityFramework usesa concept
called Convention Over Configuration
(EF has a lot of conventions built into it & based on these conventions
it creates these tables),
which work most of the time but not always!
And another problem is Data Type of nvarchar are set for for Artist_Id
and not integer is because in the
Application User Table the primary key is of nvachar of 128, that's how
ASP.NET is designed, Good or Bad?

Solution: Overriding Convetions by using Data Annotaions or Fluent API.
Will use Data Annotations for now
& later refactor the code and use Fluent API!

Add-Migration OverrideConventionsForGigsAndGenres  - We ran this to fix
our problem with null strings but nvarchar(ASP.NET EF thing) stays the
same but not null