SQL Script to populate Genres Table using a Blank Migration
Created a blank migration and used that to execute the SQL script. We
get an empty migration when there is NO changes to the domain model.
Wrote SQL script on the Up() method aswell as corresponding down()
method

The migration popoulated the Genres Table, a lot simpler and faster than
creating an admin page to populate a list of genres, plus in the future
it's easier to populate more genres or delete by using a SQL Script