Bootstrap Form + ViewModel (Presentation Model)
Built a Form using Bootstrap

PROBLEM:
Our Gig Model has a property DateTime which captures both date and time
BUT typically (for usability reasons) u should have 2 separate fields,
one for date and one for time.

SOLUTION: ViewModels / Presentation Model - What we are going to present
to the user is different from our domain model because we have DateTime
as single property but we will separate this on the view!
ViewModel  or PresentationModel is NOT part of the domain! It is a
different class purely used for presentation!
Created ViewModels folder, then a class called GigFormViewModel inside
it. Then added separate properties for Date and Time unlike the DateTime
property we have in Gigs.cs class

Changed DateTime to Date property and Time to Time property  in the view
where we have the form (Create.cshtml)