Added Genre dropdown list to Form
Created Genres property and set it as integer because dropdowns accept
integers and it will be stored as GenreId in the database!

Not worried about Dependancy Injection or Repositories at this point as
you can see because controller is dealing with DAL - Not good SOLID
practice! - Will change it later