Saving Data From Form
Saving Data From Form - No validation at this point, assuming everything
is validated from user!

So we have a Gig object and we need to add it to our _context so it can
be tracked by EF and we call save changes() method and EF will generate
a SQL statement and execute it against the database.

Temporarily, we are going to redirect user to the Home Page and later in
the next Use Cases i will replace the Home page with a list of all the
upcoming gigs!